feat: dup plugin

+++ moe/plugins/remove/rm_core.py
@@ -2,8 +2,14 @@
 
 import logging
 
+import sqlalchemy
+import sqlalchemy.exc
+from sqlalchemy.orm.session import Session
+
 from moe.config import Config, MoeSession
+from moe.library.extra import Extra
 from moe.library.lib_item import LibItem
+from moe.library.track import Track
 
 __all__ = ["remove_item"]
 
@@ -13,9 +19,25 @@ log = logging.getLogger("moe.remove")
 def remove_item(config: Config, item: LibItem):
     """Removes an item from the library."""
     log.debug(f"Removing item from the library. [{item=!r}]")
-
     session = MoeSession()
-    session.delete(item)
-    session.flush()
+
+    insp = sqlalchemy.inspect(item)
+    if insp.persistent:
+        session.delete(item)
+    elif insp.pending:
+        session.expunge(item)
+        if isinstance(item, (Track, Extra)):
+            item.album_obj = None  # type: ignore
+
+    try:
+        session.flush()
+    except sqlalchemy.exc.InvalidRequestError:
+        # session is currently flushing - delete in separate session to ensure it
+        # occurs before any inserts or updates in the original session
+        if insp.persistent:
+            session.expunge(item)
+            new_session = Session(session.connection())
+            new_session.delete(item)
+            new_session.flush()
 
     log.info(f"Removed item from the library. [{item=!r}]")
